auto-dbt-docs/pyproject.toml
[project]
name = "dbt-docs-generator"
version = "0.1.0"
description = "Automated dbt documentation generator for Mintlify"
dependencies = [
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "Jinja2>=3.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]